+++ lib/live_set/live_set_all_tracks.rb
@@ -16,6 +16,23 @@ class AllTracks
     end
   end
 
+  def ready_for_p3s
+    @track_instances.all?(&:clips_collected?) && all_frozen
+  end
+
+  def p3s_tx_msg
+    return 'Ready for transfer to Push 3 Standalone'.green if ready_for_p3s
+
+    response = []
+    response << 'Some tracks are not frozen'.red unless all_frozen
+    @track_instances.each do |track|
+      track.audio_clips.each do |clip|
+        response << "Clip '#{clip.relative_path}' has not been collected".red unless clip.relative_path.start_with?('Samples/Imported/')
+      end
+    end
+    response.join "\n"
+  end
+
   def summary
     all_frozen_msg = all_frozen ? ' (All are frozen)'.yellow : ''
     push_warning = all_frozen ? '' : "\nWarning: some tracks are not frozen, so this set should not be transferred to Push 3 Standalone.".red
@@ -27,6 +44,7 @@ class AllTracks
                    end
     total_set_size = "Total set size: #{human_file_size total_size}".yellow
     <<~END_MSG
+      #{p3s_tx_msg}
       #{total_set_size}#{size_warning}#{push_warning}
       #{@track_instances.count} tracks#{all_frozen_msg}:
     END_MSG